+++ lib/src/analyzer_plugin/analyzer_plugin.dart
@@ -31,7 +31,7 @@ class MetricsAnalyzerPlugin extends ServerPlugin {
   String get name => 'Dart Code Metrics';
 
   @override
-  String get version => '1.4.0';
+  String get version => '1.5.0';
 
   @override
   void contentChanged(String path) {