+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Unreleased
+# 1.5.0
 - Add experimental static code diagnostics double-literal-format, no-boolean-literal-compare
 
 # 1.4.0